Fix symbol parsing
