+++ packages/seroval/src/tree/async.ts
@@ -27,6 +27,7 @@ import {
   createReferenceNode,
   createArrayBufferNode,
   createDataViewNode,
+  createSymbolNode,
 } from './primitives';
 import { hasReferenceID } from './reference';
 import {
@@ -34,7 +35,6 @@ import {
   getErrorOptions,
   isIterable,
 } from './shared';
-import { WellKnownSymbols } from './symbols';
 import {
   SerovalAggregateErrorNode,
   SerovalArrayNode,
@@ -548,14 +548,7 @@ async function parse<T>(
       throw new Error('Unsupported type');
     }
     case 'symbol':
-      if (hasReferenceID(current)) {
-        const id = createIndexedValue(ctx, current);
-        if (ctx.markedRefs.has(id)) {
-          return createIndexedValueNode(id);
-        }
-        return createReferenceNode(id, current);
-      }
-      return createWKSymbolNode(ctx, current as WellKnownSymbols);
+      return createSymbolNode(ctx, current);
     case 'function': {
       assert(hasReferenceID(current), 'Cannot serialize function without reference ID.');
       const id = createIndexedValue(ctx, current);
